Set X-Forwarded-Port header to fix rack middleware
Rack middleware currently generates incorrect URLs when serving an app
through pow with this module. During a request to `http://myapp.pow`,
calling `Rack::Request#url` produces `http://myapp.pow:30559`.

Rack is incorrectly using the port that pow is listening for proxied
connections on. This header tells rack what the original port that the
end user connected to was, which lets it generate the right URLs.